% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stars.R
\name{add_stars}
\alias{add_stars}
\title{Add Stars to Estimates}
\usage{
add_stars(
  estimates = NULL,
  pvalues = NULL,
  thresholds = c(0.01, 0.05, 0.1),
  digits = 3
)
}
\arguments{
\item{estimates}{Numeric vector. Coefficients.}

\item{pvalues}{Numeric vector. Corresponding p-values.}

\item{thresholds}{Numeric vector. Three cut-off points for p-values.}

\item{digits}{Integer. Number of digits for rounding.}
}
\value{
Character vector with significance levels associated with coefficients.
}
\description{
Add stars to a vector of coefficients based on the corresponding p-values and specified thresholds.
}
\author{
Nicolas Mangin
}
