% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.R
\name{as.data.frame.fcat}
\alias{as.data.frame.fcat}
\title{Convert a format catalog to a data frame}
\usage{
\method{as.data.frame}{fcat}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{The format catalog to convert.}

\item{row.names}{Row names of the return data frame.  Default is NULL.}

\item{optional}{TRUE or FALSE value indicating whether converting to
syntactic variable names is options.  In the case of formats, the 
resulting data frame will always be returned with syntactic names, and 
this parameter is ignored.}

\item{...}{Any follow-on parameters.}
}
\value{
A data frame that contains the values stored in the format 
catalog.
}
\description{
This function takes the information stored in a format 
catalog, and coverts it to the data frame.  The data frame format is 
useful for storage, editing, saving to a spreadsheet, etc.  The 
data frame show the name of the formats, their class, and the format 
expression.  For use-defined formats, the data frame populates 
additional columns for the label and order.
}
\seealso{
Other fcat: 
\code{\link{as.fcat.fmt_lst}()},
\code{\link{as.fcat.list}()},
\code{\link{as.fcat}()},
\code{\link{fcat}()},
\code{\link{is.fcat}()},
\code{\link{print.fcat}()},
\code{\link{read.fcat}()},
\code{\link{write.fcat}()}
}
\concept{fcat}
