% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.R
\name{as.fcat.data.frame}
\alias{as.fcat.data.frame}
\title{Convert a data frame to a format catalog}
\usage{
\method{as.fcat}{data.frame}(x)
}
\arguments{
\item{x}{The data frame to convert.}
}
\value{
A format catalog based on the information contained in the 
input data frame.
}
\description{
This function takes a data frame as input
and converts it to a format catalog based on the information contained
in the data frame. The data frame should have 5 columns: "Name", "Type",
"Expression", "Label" and "Order".
}
\details{
Valid values for the "Type" column are as follows:
\itemize{
\item \strong{U}: User Defined List created with the \code{value} function.
\item \strong{S}: A formatting string of formatting codes.
\item \strong{F}: A vectorized function.
\item \strong{V}: A named vector lookup.}
}
