% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.R
\name{as.fcat.fmt_lst}
\alias{as.fcat.fmt_lst}
\title{Convert a formatting list to a format catalog}
\usage{
\method{as.fcat}{fmt_lst}(x)
}
\arguments{
\item{x}{The formatting list to convert.}
}
\value{
A format catalog based on the formats contained in the input
formatting list.
}
\description{
The \code{as.fcat.list} function converts a formatting list
to a format catalog.  For additional information on formatting lists,
see \code{\link{flist}}.
}
\seealso{
Other fcat: 
\code{\link{as.data.frame.fcat}()},
\code{\link{as.fcat}()},
\code{\link{as.fcat.data.frame}()},
\code{\link{as.fcat.list}()},
\code{\link{fcat}()},
\code{\link{is.fcat}()},
\code{\link{print.fcat}()},
\code{\link{read.fcat}()},
\code{\link{write.fcat}()}
}
\concept{fcat}
