% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.R
\name{as.fcat.list}
\alias{as.fcat.list}
\title{Convert a list to a format catalog}
\usage{
\method{as.fcat}{list}(x)
}
\arguments{
\item{x}{The list to convert.  List must contained named formats.}
}
\value{
A format catalog based on the formats contained in the input list.
}
\description{
The \code{as.fcat.list} function converts a list of formats
to a format catalog.
}
\seealso{
Other fcat: 
\code{\link{as.data.frame.fcat}()},
\code{\link{as.fcat.data.frame}()},
\code{\link{as.fcat.fmt_lst}()},
\code{\link{as.fcat}()},
\code{\link{fcat}()},
\code{\link{is.fcat}()},
\code{\link{print.fcat}()},
\code{\link{read.fcat}()},
\code{\link{write.fcat}()}
}
\concept{fcat}
