% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.R
\name{as.fcat}
\alias{as.fcat}
\title{Generic Casting Method for Format Catalogs}
\usage{
as.fcat(x)
}
\arguments{
\item{x}{The object to cast.}
}
\value{
A formatting object, created using the information in the 
input object.
}
\description{
A generic method for casting objects to
a format catalog.  Individual objects will inherit from this function.
}
\seealso{
Other fcat: 
\code{\link{as.data.frame.fcat}()},
\code{\link{as.fcat.fmt_lst}()},
\code{\link{as.fcat.list}()},
\code{\link{fcat}()},
\code{\link{is.fcat}()},
\code{\link{print.fcat}()},
\code{\link{read.fcat}()},
\code{\link{write.fcat}()}
}
\concept{fcat}
