% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{as.fmt}
\alias{as.fmt}
\title{Generic casting method for formats}
\usage{
as.fmt(x)
}
\arguments{
\item{x}{The object to cast.}
}
\value{
A formatting object, created using the information in the 
input object.
}
\description{
A generic method for casting objects to
a format.  Individual objects will inherit from this function.
}
\seealso{
Other fmt: 
\code{\link{as.data.frame.fmt}()},
\code{\link{as.fmt.data.frame}()},
\code{\link{condition}()},
\code{\link{is.format}()},
\code{\link{labels.fmt}()},
\code{\link{print.fmt}()},
\code{\link{value}()}
}
\concept{fmt}
