% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl.R
\name{dbl}
\alias{dbl}
\title{Install scholR environment}
\usage{
dbl(x, digits = 2)
}
\arguments{
\item{x}{Double or Integer. Number to be properly formatted.}

\item{digits}{Integer. Number of digits.}
}
\value{
Character. Formatted number.
}
\description{
dbl format as a number with two digits.
}
\author{
Nicolas Mangin
}
