% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcat.R
\name{fcat}
\alias{fcat}
\title{Create a format catalog}
\usage{
fcat(...)
}
\arguments{
\item{...}{A set of formats. Pass the formats as a name/value pair.  Multiple
name/value pairs may be separated by a comma.}
}
\value{
The format catalog object.
}
\description{
A format catalog is a collection of formats.  This format
collection allows you to manage and store formats as a unit.  The 
\code{fcat} function defines the format catalog.
}
\details{
A format catalog is an S3 object of class "fcat".  The purpose of 
the catalog is to combine related formats, and allow you to manipulate all
of them as a single object.  The format catalog can be saved to a file 
using the \code{\link{write.fcat}} function.  A format catalog can also
be converted to/from a data frame using the \code{\link{as.fcat.data.frame}}
and \code{\link{as.data.frame.fcat}} functions.
}
\examples{
# Create format catalog
c1 <- fcat(num_fmt  = "\%.1f",
           label_fmt = value(condition(x == "A", "Label A"),
                             condition(x == "B", "Label B"),
                             condition(TRUE, "Other")),
           date_fmt = "\%d\%b\%Y")

# Use formats in the catalog
fapply(2, c1$num_fmt)
fapply(c("A", "B", "C", "B"), c1$label_fmt)
fapply(Sys.Date(), c1$date_fmt)
}
\seealso{
Other fcat: 
\code{\link{as.data.frame.fcat}()},
\code{\link{as.fcat.fmt_lst}()},
\code{\link{as.fcat.list}()},
\code{\link{as.fcat}()},
\code{\link{is.fcat}()},
\code{\link{print.fcat}()},
\code{\link{read.fcat}()},
\code{\link{write.fcat}()}
}
\concept{fcat}
