% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_cnt_pct}
\alias{fmt_cnt_pct}
\title{Formatted count and percent}
\usage{
fmt_cnt_pct(x, denom = NULL, format = "\%5.1f")
}
\arguments{
\item{x}{The input data vector.}

\item{denom}{The denominator to use for the percentage. By default, the 
parameter is NULL, meaning the function will use non-missing values of the 
data vector as the denominator.  Otherwise, supply the denominator as 
a numeric value.}

\item{format}{A formatting string suitable for input into the 
\code{\link[base]{sprintf}} function.  By default, this format is
defined as "\%5.1f", which displays the value with one decimal place.}
}
\value{
The formatted median value
}
\description{
A function to calculate and format a count and percent
}
\seealso{
Other helpers: 
\code{\link{fmt_mean_sd}()},
\code{\link{fmt_median}()},
\code{\link{fmt_n}()},
\code{\link{fmt_quantile_range}()},
\code{\link{fmt_range}()}
}
\concept{helpers}
