% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_mean_sd}
\alias{fmt_mean_sd}
\title{Formatted mean and standard deviation}
\usage{
fmt_mean_sd(x, format = "\%.1f")
}
\arguments{
\item{x}{The input data vector.}

\item{format}{A formatting string suitable for input into the 
\code{\link[base]{sprintf}} function.  By default, this format is
defined as "\%.1f", which displays the mean and standard deviation with 
one decimal place.}
}
\value{
The formatted mean and standard deviation
}
\description{
A function to calculate and format a mean and standard deviation
}
\seealso{
Other helpers: 
\code{\link{fmt_cnt_pct}()},
\code{\link{fmt_median}()},
\code{\link{fmt_n}()},
\code{\link{fmt_quantile_range}()},
\code{\link{fmt_range}()}
}
\concept{helpers}
