% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_n}
\alias{fmt_n}
\title{Formatted Count}
\usage{
fmt_n(x)
}
\arguments{
\item{x}{The input data vector.}
}
\value{
The formatted count value
}
\description{
A function to calculate and format a numeric count
}
\seealso{
Other helpers: 
\code{\link{fmt_cnt_pct}()},
\code{\link{fmt_mean_sd}()},
\code{\link{fmt_median}()},
\code{\link{fmt_quantile_range}()},
\code{\link{fmt_range}()}
}
\concept{helpers}
