% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_quantile_range}
\alias{fmt_quantile_range}
\title{Formatted Quantile Range}
\usage{
fmt_quantile_range(x, format = "\%.1f", sep = "-")
}
\arguments{
\item{x}{The input data vector.}

\item{format}{A formatting string suitable for input into the 
\code{\link[base]{sprintf}} function.  By default, this format is
defined as "\%.1f", which displays the value with one decimal place.}

\item{sep}{The character to use as a separator between the two quantiles.}
}
\value{
The formatted quantile range
}
\description{
A function to calculate and format a quantile range
}
\seealso{
Other helpers: 
\code{\link{fmt_cnt_pct}()},
\code{\link{fmt_mean_sd}()},
\code{\link{fmt_median}()},
\code{\link{fmt_n}()},
\code{\link{fmt_range}()}
}
\concept{helpers}
