% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_range}
\alias{fmt_range}
\title{Formatted Range}
\usage{
fmt_range(x, format = "\%s", sep = "-")
}
\arguments{
\item{x}{The input data vector.}

\item{format}{A formatting string suitable for input into the 
\code{\link[base]{sprintf}} function.  By default, this format is
defined as "\%s", which simply converts the value to a string with no
specific formatting.}

\item{sep}{The token used to separate the minimum and maximum range
values.  Default value is a hyphen ("-").}
}
\value{
The formatted range values
}
\description{
A function to calculate and format a numeric range
}
\seealso{
Other helpers: 
\code{\link{fmt_cnt_pct}()},
\code{\link{fmt_mean_sd}()},
\code{\link{fmt_median}()},
\code{\link{fmt_n}()},
\code{\link{fmt_quantile_range}()}
}
\concept{helpers}
