% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtr.R
\name{FormattingCodes}
\alias{FormattingCodes}
\title{Formatting Strings}
\description{
Formatting codes for formatting strings follow the conventions 
of the base R \code{\link[base]{format}} and \code{\link[base]{sprintf}}
functions. See below for further details.
}
\details{
Here is a table:
|Code|Description|Example|
|----|-----------|-------|
|%%d |Date.      |"%%d%%m%%Y" = "01/01/1970"|

That was a table.
}
