% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{int}
\alias{int}
\title{Install scholR environment}
\usage{
int(x, tie = "up")
}
\arguments{
\item{x}{Double or Integer. Number to be properly formatted.}

\item{tie}{Character. Rounding 0.5, "up", "down", or to "even".}
}
\value{
Character. Formatted number.
}
\description{
int formats as an integer.
}
\author{
Nicolas Mangin
}
