% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist.R
\name{is.flist}
\alias{is.flist}
\title{Is object a formatting list}
\usage{
is.flist(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
TRUE or FALSE, depending on class of object.
}
\description{
Determines if object is a formatting list of class 'fmt_lst'.
}
\examples{
# Create flist
flst <- flist("\%d\%b\%Y", "\%.1f")
is.flist(flst)
is.flist("A")
}
\seealso{
Other flist: 
\code{\link{as.data.frame.fmt_lst}()},
\code{\link{as.flist}()},
\code{\link{as.flist.data.frame}()},
\code{\link{as.flist.fcat}()},
\code{\link{as.flist.list}()},
\code{\link{as.flist.tbl_df}()},
\code{\link{flist}()},
\code{\link{print.fmt_lst}()},
\code{\link{read.flist}()},
\code{\link{write.flist}()}
}
\concept{flist}
