% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct.R
\name{pct}
\alias{pct}
\title{Install scholR environment}
\usage{
pct(x, digits = 2)
}
\arguments{
\item{x}{Double or Integer. Number to be properly formatted.}

\item{digits}{Integer. Number of digits.}
}
\value{
Character. Formated number.
}
\description{
pct transforms numbers between 0 and 1 in percentages.
}
\author{
Nicolas Mangin
}
