% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{print.fmt}
\alias{print.fmt}
\title{Print a format}
\usage{
\method{print}{fmt}(
  x,
  ...,
  name = deparse(substitute(x, env = environment())),
  verbose = FALSE
)
}
\arguments{
\item{x}{An object of class "fmt".}

\item{...}{Any follow-on parameters to the print function.}

\item{name}{The name of the format to print. By default, the variable
name that holds the format will be used.}

\item{verbose}{Turn on or off verbose printing mode.  Verbose mode will
print object as a list.  Otherwise, the object will be printed as a table.}
}
\description{
Prints a format object.  This function is
a class-specific implementation of the the generic \code{print} method.
}
\seealso{
Other fmt: 
\code{\link{as.data.frame.fmt}()},
\code{\link{as.fmt}()},
\code{\link{as.fmt.data.frame}()},
\code{\link{condition}()},
\code{\link{is.format}()},
\code{\link{labels.fmt}()},
\code{\link{value}()}
}
\concept{fmt}
