% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flist.R
\name{print.fmt_lst}
\alias{print.fmt_lst}
\title{Print a formatting list}
\usage{
\method{print}{fmt_lst}(x, ..., verbose = FALSE)
}
\arguments{
\item{x}{The formatting list to print}

\item{...}{Follow-on parameters to the print function}

\item{verbose}{Whether to print in summary or list-style.}
}
\description{
Print a formatting list
}
\seealso{
Other flist: 
\code{\link{as.data.frame.fmt_lst}()},
\code{\link{as.flist}()},
\code{\link{as.flist.data.frame}()},
\code{\link{as.flist.fcat}()},
\code{\link{as.flist.list}()},
\code{\link{as.flist.tbl_df}()},
\code{\link{flist}()},
\code{\link{is.flist}()},
\code{\link{read.flist}()},
\code{\link{write.flist}()}
}
\concept{flist}
