% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_estimate.R
\name{report_estimate}
\alias{report_estimate}
\title{Format Estimated Coefficients for Reporting}
\usage{
report_estimate(
  symbol = "$\\beta$",
  estimate = NULL,
  std_error = NULL,
  threshold = 0.05,
  digits = 3,
  statistics = c("cf", "ci", "pv")
)
}
\arguments{
\item{symbol}{Character. Name or identifier of the coefficient.}

\item{estimate}{Numeric. Estimate of the coefficient.}

\item{std_error}{Numeric. Standard error around the estimate.}

\item{threshold}{Numeric. Probability of 0 contained in the confidence interval (e.g. 0.05 for a 95\% confidence interval).}

\item{digits}{Integer. Number of digits for rounding.}

\item{statistics}{Character vector. Statistics to report: "cf" for the coefficient, "ci" for the confidence internal, "pv" for the p-value.}
}
\value{
Character with specified statistics between parentheses.
}
\description{
Format in-text the reporting of confidence intervals or p-values around an estimate.
}
\author{
Nicolas Mangin
}
