% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{sci}
\alias{sci}
\title{Install scholR environment}
\usage{
sci(x, digits = 2)
}
\arguments{
\item{x}{Double or Integer. Number to be properly formatted.}

\item{digits}{Integer. Number of digits.}
}
\value{
Character. Number under scientific notation.
}
\description{
sci format scientifically.
}
\author{
Nicolas Mangin
}
