% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{as.fmt}
\alias{as.fmt}
\title{Generic casting method for formats}
\usage{
as.fmt(x)
}
\arguments{
\item{x}{The object to cast.}
}
\value{
A formatting object, created using the information in the 
input object.
}
\description{
A generic method for casting objects to
a format.  Individual objects will inherit from this function.
}
