% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattr.R
\name{fattr}
\alias{fattr}
\title{Set formatting attributes}
\usage{
fattr(x, format = NULL, width = NULL, justify = NULL)
}
\arguments{
\item{x}{The vector to assign attributes to.}

\item{format}{The format to assign to the format attribute.  Format 
can be a formatting string, a named vector decode, a vectorized
formatting function, or a formatting list.}

\item{width}{The desired width of the formatted output.}

\item{justify}{Justification of the output vector. Valid values are 
'none', 'left', 'right', 'center', or 'centre'.}
}
\description{
Assign formatting attributes to a vector
}
\details{
The \code{fattr} function is a convenience function for assigning 
formatting attributes to a vector.  The function accepts parameters
for format, width, and justify.  Any formatting attributes assigned 
to a vector can be applied using \code{\link{fapply}} or
\code{\link{format_data}}.
}
\examples{
# Create vector
a <- c(1.3243, 5.9783, 2.3848)

# Assign format attriutes
a <- fattr(a, format = "\%.1f", width = 10, justify = "right")

# Apply format attributes
fapply(a)
}
\seealso{
\code{\link{format_data}} to apply formats to a data frame, 
 \code{\link{fapply}} to apply formats to a vector.
}
