% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_median}
\alias{fmt_median}
\title{Formatted Median}
\usage{
fmt_median(x, format = "\%.1f")
}
\arguments{
\item{x}{The input data vector.}

\item{format}{A formatting string suitable for input into the 
\code{\link[base]{sprintf}} function.  By default, this format is
defined as "\%.1f", which displays the value with one decimal place.}
}
\value{
The formatted median value
}
\description{
A function to calculate and format a median
}
\seealso{
Other helpers: 
\code{\link{fmt_cnt_pct}()},
\code{\link{fmt_mean_sd}()},
\code{\link{fmt_n}()},
\code{\link{fmt_quantile_range}()},
\code{\link{fmt_range}()}
}
\concept{helpers}
