% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtr.R
\docType{package}
\name{fmtr}
\alias{fmtr}
\title{Formatting data frames}
\description{
The \strong{fmtr} package contains functions to 
format data frames.
}
\details{
The \strong{fmtr} package helps format data frames, similar 
to the way datasets are formatted in SAS®.  Formats are assigned
to columns using a \strong{format} attribute.  Formats are then applied
by calling the \code{format()} function on the data frame.  
The \code{format()}
function returns a new data frame with the formats applied to each column
as specified. 

All functions in the \code{fmtr} package also work with tidyverse tibbles.

The \strong{fmtr} package also contains a function named 
\code{value} for defining user-defined formats in a manner
similar to SAS® formats.
}
\section{Functions}{

The three main functions included in the \strong{fmtr} package are 
as follows:
\itemize{
  \item \code{\link{fdata}}
  \item \code{\link{fapply}}
  \item \code{\link{value}}
}

The \code{fdata()} function applies formatting 
to dataframe.  

#' The \code{fapply()} function applies the formatting 
to a vector.  

The \code{value()} function creates a new
user-defined format.
}

\section{Available Formats}{

The formats that can be used with \strong{fmtr} include the following:
\itemize{
  \item Named vectors
  \item Vectorized functions
  \item User-defined formats
  \item Format lists
}

Named vectors map one string value to another string value.

Vectorized functions can be those created by the user or by a formatting
function like those created with Base R or the \strong{scales} package.

User-defined function are those created by the \strong{fmtr} 
\code{value} function.  See \code{link{value()}} for 
additional details.

Format lists are lists which contain any of the above format types.
The format lists allow the user to apply one of several formats to the 
column.  See \code{link{fapply()}} for 
additional details.


See function documentation for additional details.
}

