% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justification.R
\name{justification}
\alias{justification}
\alias{"justification<-"}
\title{Get or set justification for data frame columns}
\usage{
justification(x)
}
\arguments{
\item{x}{A data frame or tibble.}
}
\value{
A named list of justification values.
}
\description{
The \code{justification} function extracts all assigned justify
attributes from a 
data frame, and returns them in a named list. The function also
assigns justify attributes from a named list.
}
\details{
If justification values are assigned to the "justify" 
attributes of the data frame
columns, the \code{justification} function will extract those values.  The 
function will return the justification values in a named list, 
where the names
correspond to the name of the column that the justification was assigned to.
If a column does not have a justify attribute assigned, that column
will not be included in the list. 

When used on the receiving side of an assignment, the function will assign
justification to a data frame.  The justification values 
should be in a named list, where
each name corresponds to the name of the data frame column to assign
values to.
}
\examples{
# Take subset of data
df1 <- mtcars[1:10, c("mpg", "cyl") ]

# Print current state
print(df1)

# Assign justification
justification(df1) <- list(mpg = "left", cyl = "center")
widths(df1) <- list(mpg = 12, cyl = 10)

format(df1)


# Display justification
justification(df1)
widths(df1)
}
\seealso{
\code{\link{format_data}} to display formatted data, 
\code{\link{value}} to create user-defined formats, and 
\code{\link{fapply}} to apply formatting to a vector.
}
